Added live config support for VideoCapture apps
