+++ documentation/src/utils/schemas/videoCaptureConfig.schema.ts
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+import { CaptureWorkflow } from '@monkvision/types';
+import { SharedCaptureAppConfigSchema } from '@site/src/utils/schemas/sharedConfig.schema';
+
+export const VideoCaptureAppConfigSchema = z
+  .object({
+    workflow: z.literal(CaptureWorkflow.VIDEO),
+  })
+  .and(SharedCaptureAppConfigSchema);
