+++ packages/common/src/apps/appState.ts
@@ -1,21 +1,22 @@
-import { CaptureAppConfig, Sight, SteeringWheelPosition, VehicleType } from '@monkvision/types';
+import {
+  PhotoCaptureAppConfig,
+  Sight,
+  SteeringWheelPosition,
+  VehicleType,
+  VideoCaptureAppConfig,
+} from '@monkvision/types';
 import { createContext } from 'react';
 import { LoadingState } from '../hooks';
 
 /**
- * Application state usually used by Monk applications to configure and handle the current user journey.
+ * Shared app states values by both photo and video capture workflows.
  */
-export interface MonkAppState {
+export interface SharedMonkAppState {
   /**
    * LoadingState indicating if the application state is loading. If it is loading it usually means that the provider
    * did not have time to fetch the parameter values.
    */
   loading: LoadingState;
-  /**
-   * The current configuration of the application.
-   */
-  config: CaptureAppConfig;
-
   /**
    * The authentication token representing the currently logged-in user. If this param is `null`, it means the user is
    * not logged in.
@@ -26,6 +27,24 @@ export interface MonkAppState {
    * param is `null`, it probably means that the inspection must be created by the app.
    */
   inspectionId: string | null;
+  /**
+   * Setter function used to set the current auth token.
+   */
+  setAuthToken: (value: string | null) => void;
+  /**
+   * Setter function used to set the current inspection ID.
+   */
+  setInspectionId: (value: string | null) => void;
+}
+
+/**
+ * App state values available in PhotoCapture applications.
+ */
+export interface PhotoCaptureAppState extends SharedMonkAppState {
+  /**
+   * The current configuration of the application.
+   */
+  config: PhotoCaptureAppConfig;
   /**
    * The current vehicle type of the app. This value usually helps to choose which sights to display to the user, or
    * which car 360 wireframes to use for the inspection report.
@@ -43,15 +62,6 @@ export interface MonkAppState {
    * Getter function used to get the current Sights based on the current VehicleType, SteeringWheel position etc.
    */
   getCurrentSights: () => Sight[];
-
-  /**
-   * Setter function used to set the current auth token.
-   */
-  setAuthToken: (value: string | null) => void;
-  /**
-   * Setter function used to set the current inspection ID.
-   */
-  setInspectionId: (value: string | null) => void;
   /**
    * Setter function used to set the current vehicle type.
    */
@@ -61,6 +71,21 @@ export interface MonkAppState {
    */
   setSteeringWheel: (value: SteeringWheelPosition | null) => void;
 }
+
+/**
+ * App state values available in PhotoCapture applications.
+ */
+export interface VideoCaptureAppState extends SharedMonkAppState {
+  /**
+   * The current configuration of the application.
+   */
+  config: VideoCaptureAppConfig;
+}
+
+/**
+ * Application state usually used by Monk applications to configure and handle the current user journey.
+ */
+export type MonkAppState = PhotoCaptureAppState | VideoCaptureAppState;
 
 /**
  * React context used to store the current Monk application state.
